refactor: separate change stream prose tests
